change data type passed to jparepository

+++ tech-m-obras/tech-m-obras/src/main/java/com/techmobras/techmobras/repository/ObraDetalhesTecnicosRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.techmobras.techmobras.model.ObraDetalhesTecnicos;
 
-public interface ObraDetalhesTecnicosRepository extends JpaRepository<ObraDetalhesTecnicos, Long>{
+public interface ObraDetalhesTecnicosRepository extends JpaRepository<ObraDetalhesTecnicos, Integer>{
     
 }
