+++ tech-m-obras/tech-m-obras/src/main/java/com/techmobras/techmobras/repository/ObraLocalizacaoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.techmobras.techmobras.model.ObraLocalizacao;
 
-public interface ObraLocalizacaoRepository extends JpaRepository<ObraLocalizacao, Long>{
+public interface ObraLocalizacaoRepository extends JpaRepository<ObraLocalizacao, Integer>{
     
 }
